test: update tests
move fixtures into its own file
